support for grove RGB LCD 16x2
