+++ devs/samples/testrig.ts
@@ -2,7 +2,7 @@ import * as ds from "@devicescript/core"
 import {
     SSD1306Driver,
     startBME680,
-    startCharacterScreen,
+    startCharacterScreenDisplay,
 } from "@devicescript/drivers"
 import { fetch } from "@devicescript/net"
 import {
@@ -27,7 +27,7 @@ const servo = startServo({ pin: pins.A2 })
 const potentiometer = startPotentiometer({ pin: pins.A0 })
 const buzzer = startBuzzer({ pin: pins.A1 })
 const { temperature, humidity, pressure } = await startBME680()
-const display = await startCharacterScreen(
+const display = await startCharacterScreenDisplay(
     new SSD1306Driver({ width: 64, height: 48 })
 )
 const btnA = new ds.Button()
