+++ packages/drivers/src/index.ts
@@ -8,7 +8,7 @@ export * from "./aht20"
 export * from "./ltr390"
 export * from "./bme680"
 export * from "./ssd1306"
-export * from "./characterscreen"
+export * from "./characterscreendisplay"
 export * from "./indexedscreen"
 export * from "./dotmatrix"
 export * from "./st7735"
@@ -18,6 +18,7 @@ export * from "./ledserver"
 export * from "./accelerometer"
 export * from "./sh110x"
 export * from "./da213b"
+export * from "./grovergblcd"
 
 export * from "./esp32c3fh4rgb"
 export * from "./picobricks"
