+++ packages/sampleprj/src/mainwroomssd1306.ts
@@ -1,5 +1,5 @@
 // https://github.com/microsoft/devicescript/issues/559
-import { SSD1306Driver, startCharacterScreen } from "@devicescript/drivers";
+import { SSD1306Driver, startCharacterScreenDisplay } from "@devicescript/drivers";
 import { configureHardware } from "@devicescript/servers";
 // https://microsoft.github.io/devicescript/devices/esp32/esp32-devkit-c
 import { pins } from "@dsboard/esp32_devkit_c";
@@ -15,5 +15,5 @@ configureHardware({
 // double check i2c addr
 const ssdDisplay = new SSD1306Driver({ height: 64, width: 128, devAddr: 0x3d })
 
-const characterScreen = await startCharacterScreen(ssdDisplay)
+const characterScreen = await startCharacterScreenDisplay(ssdDisplay)
 await characterScreen.message.write("Hello world!")
