+++ vscode/src/server-info.json
@@ -254,7 +254,7 @@
         },
         {
             "label": "Sensirion SHTC3",
-            "detail": "Start driver for Sensirion SHTC3 temperature/humidity sensor at I2C `0x70`.",
+            "detail": "Start driver for Sensirion SHTC3 temperature/humidity sensor at I2C `0x70`.\r",
             "startName": "startSHTC3",
             "classIdentifiers": [
                 337754823,
@@ -267,7 +267,7 @@
         },
         {
             "label": "Sensirion SHT30",
-            "detail": "Start driver for Sensirion SHT30 temperature/humidity sensor at I2C `0x44` or `0x45` (default is `0x44`)",
+            "detail": "Start driver for Sensirion SHT30 temperature/humidity sensor at I2C `0x44` or `0x45` (default is `0x44`)\r",
             "startName": "startSHT30",
             "classIdentifiers": [
                 337754823,
@@ -280,7 +280,7 @@
         },
         {
             "label": "AHT20",
-            "detail": "Start driver for AHT20 temperature/humidity sensor at I2C address `0x38`.",
+            "detail": "Start driver for AHT20 temperature/humidity sensor at I2C address `0x38`.\r",
             "startName": "startAHT20",
             "classIdentifiers": [
                 337754823,
@@ -293,7 +293,7 @@
         },
         {
             "label": "LITEON LTR-390UV-01",
-            "detail": "Start driver for LITEON LTR-390UV-01 UV/ambient light sensor at I2C address `0x53`.",
+            "detail": "Start driver for LITEON LTR-390UV-01 UV/ambient light sensor at I2C address `0x53`.\r",
             "startName": "startLTR390",
             "classIdentifiers": [
                 527306128,
@@ -306,7 +306,7 @@
         },
         {
             "label": "Bosch BME68",
-            "detail": "Start driver for Bosch BME680 temperature/humidity/pressure/gas sensor at I2C `0x76` (default) or `0x77`.",
+            "detail": "Start driver for Bosch BME680 temperature/humidity/pressure/gas sensor at I2C `0x76` (default) or `0x77`.\r",
             "startName": "startBME680",
             "classIdentifiers": [
                 337754823,
@@ -321,7 +321,7 @@
         },
         {
             "label": "Pico Bricks for Raspberry Pi Pico",
-            "detail": "Drivers for the {@link https://shop.robotistan.com/products/pico-bricks | Pico Bricks } shield for Raspberry Pi Pico ({@link https://github.com/Robotistan/PicoBricks/tree/main/Documents | datasheets}).",
+            "detail": "Drivers for the {@link https://shop.robotistan.com/products/pico-bricks | Pico Bricks } shield for Raspberry Pi Pico ({@link https://github.com/Robotistan/PicoBricks/tree/main/Documents | datasheets}).\r",
             "startName": "PicoBricks",
             "imports": {
                 "PicoBricks": "@devicescript/drivers"
@@ -330,7 +330,7 @@
         },
         {
             "label": "Seeed Studio XIAO Expansion Board",
-            "detail": "Drivers for the {@link https://wiki.seeedstudio.com/Seeeduino-XIAO-Expansion-Board/ | Seeed Studio XIAO Expansion Board } for Raspberry Pi Pico.",
+            "detail": "Drivers for the {@link https://wiki.seeedstudio.com/Seeeduino-XIAO-Expansion-Board/ | Seeed Studio XIAO Expansion Board } for Raspberry Pi Pico.\r",
             "startName": "XiaoExpansionBoard",
             "imports": {
                 "XiaoExpansionBoard": "@devicescript/drivers"
@@ -339,7 +339,7 @@
         },
         {
             "label": "WaveShare Pico LCD114 for Raspberry Pi Pico",
-            "detail": "Driver for WaveShare Pico-LCD-1.14 inch.",
+            "detail": "Driver for WaveShare Pico-LCD-1.14 inch.\r",
             "startName": "WaveSharePicoLCD114",
             "imports": {
                 "WaveSharePicoLCD114": "@devicescript/drivers"
@@ -348,7 +348,7 @@
         },
         {
             "label": "Pimoroni Pico Badger",
-            "detail": "Support for Badger 2040 W (Pico W Aboard)",
+            "detail": "Support for Badger 2040 W (Pico W Aboard)\r",
             "startName": "PimoroniBadger2040W",
             "imports": {
                 "PimoroniBadger2040W": "@devicescript/drivers"
@@ -357,7 +357,7 @@
         },
         {
             "label": "KittenBot Grape:bit ESP32-C3",
-            "detail": "Support for KittenBot Grape:bit ESP32-C3",
+            "detail": "Support for KittenBot Grape:bit ESP32-C3\r",
             "startName": "KittenBotGrapeBit",
             "imports": {
                 "KittenBotGrapeBit": "@devicescript/drivers"
